Machine Learning August 2024/Assignments/a2-regression-mlflow-traefik/code/pyproject.toml
[tool.poetry]
name = "code"
version = "0.1.0"
description = "FastAPI template for ML models"
authors = ["Ulugbek Shernazarov"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.17.4"
fastapi = "^0.73.0"
pydantic = "^1.9.0"
requests = "^2.27.1"
joblib = "^1.4.2"
python-multipart = "^0.0.9"
pandas = "1.5.3"
shap = "^0.46.0"
numpy = "1.26.4"
scikit-learn = "1.3.2"
setuptools = "^59.5.0"
virtualenv = "^20.13.1"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.2"
mlflow = "^2.16.0"

[tool.poetry.dev-dependencies]
ipython = "^8.0.1"
ipdb = "^0.13.9"
flake8 = "^4.0.1"
black = "^22.1.0"
isort = "^5.10.1"
pytest = "^7.0.1"
tox = "^3.24.5"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "hug"
line_length = 120
src_paths = ["app", "tests"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
